requirements.txt
Django==4.1.1
djangorestframework~=3.13.1
django-cors-headers
djangorestframework-simplejwt
django-cryptography==1.1
django-environ==0.9.0
uWSGI==2.0.20
django-filter
requests~=2.30.0
django-search-views
django-cors-headers
django-environ==0.9.0